+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file. This project adheres
 to [Semantic Versioning](http://semver.org/). For change log format, use [Keep a Changelog](http://keepachangelog.com/).
 
+## [v2.0.0-beta3] - In progress
+
+### Added
+
+- Prefix path option for cache generator
+
+### Changed
+
+- Use `FilesystemOperator` interface instead of `Filesystem` class
+
 ## [v2.0.0-beta2] - 2022-03-21
 
 ### Changed